refactor: split quick switch shortcut settings
