+++ change-logs/2026/04/15/feature-tasks-quick-switch.md
@@ -1,1 +1,1 @@
-Added Tasks Quick Switch, a recent-task modal that cycles through active tasks across projects and opens the selected task on modifier release. Global Settings now let you choose compact task-type chips for built-in statuses and custom columns, plus either a preset shortcut or a recorded custom key chord, and the feature includes a discovery tip. Follow-up work kept the modal pinned to the real viewport via a portal, made the switcher trust live in-app settings changes immediately instead of briefly snapping back to stale backend settings on first use, restored scroll/focus to the Custom shortcut section after closing the shortcut recorder modal, strengthened the selected quick-switch row background so it reads clearly in both themes, blocked background app shortcuts while the custom shortcut modal is open, kept completed and cancelled filters visible but unselected by default, extracted the quick-switch runtime into a dedicated hook so App.tsx no longer owns the whole feature lifecycle, fixed same-tick settings changes so updated task-type filters affect the very next quick switch without requiring a restart or an extra render, gave quick switch its own non-todo task feed so selected Completed, Cancelled, and Has Questions items actually appear immediately instead of being filtered out by the backend cache path, capped the dialog to the viewport with an internal task-list scroll once the switcher grows past roughly six items, and now auto-scrolls the internal list so the selected task stays visible while cycling forward or backward with a leaner selected-row ref instead of per-item ref bookkeeping. Finally, wired the Shift-to-reverse chord so adding Shift to the configured shortcut (e.g. Shift+Alt+Tab when Alt+Tab is set) now actually cycles backward instead of being ignored by the exact-modifier matcher, matching the macOS Cmd+Tab convention. Also stopped caching the task list across opens, so an auto-transition (e.g. Your Review → Completed via a dev3 hook) no longer leaves the modal listing the task under its previous column when the next shortcut press fetches fresh data. Follow-up cleanup trimmed the tests to keep the risky keyboard, settings, and navigation paths covered without re-testing the same behavior at every layer, removed one dead quick-switch cache helper that no longer had any runtime callers, and split the bulky behavior-settings quick-switch UI into its own component so the settings section is easier to navigate and maintain.
+Added Tasks Quick Switch, a recent-task modal that cycles through active tasks across projects and opens the selected task on modifier release. Global Settings now let you choose compact task-type chips for built-in statuses and custom columns, plus either a preset shortcut or a recorded custom key chord, and the feature includes a discovery tip. Follow-up work kept the modal pinned to the real viewport via a portal, made the switcher trust live in-app settings changes immediately instead of briefly snapping back to stale backend settings on first use, restored scroll/focus to the Custom shortcut section after closing the shortcut recorder modal, strengthened the selected quick-switch row background so it reads clearly in both themes, blocked background app shortcuts while the custom shortcut modal is open, kept completed and cancelled filters visible but unselected by default, extracted the quick-switch runtime into a dedicated hook so App.tsx no longer owns the whole feature lifecycle, fixed same-tick settings changes so updated task-type filters affect the very next quick switch without requiring a restart or an extra render, gave quick switch its own non-todo task feed so selected Completed, Cancelled, and Has Questions items actually appear immediately instead of being filtered out by the backend cache path, capped the dialog to the viewport with an internal task-list scroll once the switcher grows past roughly six items, and now auto-scrolls the internal list so the selected task stays visible while cycling forward or backward with a leaner selected-row ref instead of per-item ref bookkeeping. Finally, wired the Shift-to-reverse chord so adding Shift to the configured shortcut (e.g. Shift+Alt+Tab when Alt+Tab is set) now actually cycles backward instead of being ignored by the exact-modifier matcher, matching the macOS Cmd+Tab convention. Also stopped caching the task list across opens, so an auto-transition (e.g. Your Review → Completed via a dev3 hook) no longer leaves the modal listing the task under its previous column when the next shortcut press fetches fresh data. Follow-up cleanup trimmed the tests to keep the risky keyboard, settings, and navigation paths covered without re-testing the same behavior at every layer, removed one dead quick-switch cache helper that no longer had any runtime callers, split the bulky behavior-settings quick-switch UI into its own component so the settings section is easier to navigate and maintain, and then split the shortcut controls and shortcut recorder modal into separate components so the remaining quick-switch settings file is mostly orchestration plus filter chips.
